Project/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
authx==1.4.3
cffi==1.17.1
click==8.2.1
colorama==0.4.6
cryptography==45.0.5
ecdsa==0.19.1
fastapi==0.116.1
h11==0.16.0
idna==3.10
itsdangerous==2.2.0
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
pytz==2025.2
rsa==4.9.1
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
starlette==0.47.1
thread==2.0.5
typing-inspection==0.4.1
typing_extensions==4.14.1
uvicorn==0.35.0
